Fixed a bug that prevented compilation by extracting LibrarySystem to own java-file .
